# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ᴍᴀᴅᴀʀᴀ x-ᴍᴅ WhatsApp Bot Pairing API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: pairing
    description: WhatsApp bot pairing operations
  - name: console
    description: Console log operations
  - name: contact
    description: Contact and support operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent (if account exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pairing/request:
    post:
      operationId: requestPairing
      tags: [pairing]
      summary: Request a new pairing code or QR for the bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairingRequest"
      responses:
        "200":
          description: Pairing code / QR issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pairing/status:
    get:
      operationId: getPairingStatus
      tags: [pairing]
      summary: Get current bot connection and uptime status
      responses:
        "200":
          description: Pairing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingStatus"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pairing/disconnect:
    post:
      operationId: disconnectBot
      tags: [pairing]
      summary: Disconnect the paired bot
      responses:
        "200":
          description: Bot disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pairing/stats:
    get:
      operationId: getPairingStats
      tags: [pairing]
      summary: Get pairing stats (total sessions, uptime summary, connected users)
      responses:
        "200":
          description: Pairing stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingStats"

  /console/logs:
    get:
      operationId: getConsoleLogs
      tags: [console]
      summary: Get console log entries for the current user's bot session
      responses:
        "200":
          description: Console log list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConsoleLog"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /console/clear:
    post:
      operationId: clearConsoleLogs
      tags: [console]
      summary: Clear console logs for the current user's bot session
      responses:
        "200":
          description: Logs cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact/support message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Message received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
        username:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - username
        - createdAt

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    RegisterInput:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
      required:
        - email
        - username
        - password

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
      required:
        - email

    ResetPasswordInput:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
      required:
        - token
        - password

    PairingRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          description: Phone number to pair with (with country code)
        method:
          type: string
          enum: [qr, code]
          description: Pairing method - QR scan or pairing code
      required:
        - phoneNumber
        - method

    PairingResponse:
      type: object
      properties:
        sessionId:
          type: string
        pairingCode:
          type: ["string", "null"]
          description: 8-digit pairing code (if method=code)
        qrData:
          type: ["string", "null"]
          description: QR code data string (if method=qr)
        expiresAt:
          type: string
          description: ISO timestamp when pairing expires
      required:
        - sessionId
        - pairingCode
        - qrData
        - expiresAt

    PairingStatus:
      type: object
      properties:
        connected:
          type: boolean
        phoneNumber:
          type: ["string", "null"]
        sessionId:
          type: ["string", "null"]
        connectedAt:
          type: ["string", "null"]
        uptimeSeconds:
          type: ["number", "null"]
        lastSeen:
          type: ["string", "null"]
      required:
        - connected
        - phoneNumber
        - sessionId
        - connectedAt
        - uptimeSeconds
        - lastSeen

    PairingStats:
      type: object
      properties:
        totalSessions:
          type: number
        activeSessions:
          type: number
        totalUsers:
          type: number
        averageUptimeSeconds:
          type: number
      required:
        - totalSessions
        - activeSessions
        - totalUsers
        - averageUptimeSeconds

    ConsoleLog:
      type: object
      properties:
        id:
          type: number
        level:
          type: string
          enum: [info, warn, error, success, debug]
        message:
          type: string
        timestamp:
          type: string
        sessionId:
          type: ["string", "null"]
      required:
        - id
        - level
        - message
        - timestamp
        - sessionId

    ContactInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          
        subject:
          type: string
          minLength: 3
        message:
          type: string
          minLength: 10
      required:
        - name
        - email
        - subject
        - message