# lib/api-zod/src/generated/types/consoleLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ᴍᴀᴅᴀʀᴀ x-ᴍᴅ WhatsApp Bot Pairing API
 * OpenAPI spec version: 0.1.0
 */
import type { ConsoleLogLevel } from './consoleLogLevel';

export interface ConsoleLog {
  id: number;
  level: ConsoleLogLevel;
  message: string;
  timestamp: string;
  /** @nullable */
  sessionId: string | null;
}