# lib/api-zod/src/generated/types/pairingStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ᴍᴀᴅᴀʀᴀ x-ᴍᴅ WhatsApp Bot Pairing API
 * OpenAPI spec version: 0.1.0
 */

export interface PairingStats {
  totalSessions: number;
  activeSessions: number;
  totalUsers: number;
  averageUptimeSeconds: number;
}