# lib/api-zod/src/generated/types/pairingStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ᴍᴀᴅᴀʀᴀ x-ᴍᴅ WhatsApp Bot Pairing API
 * OpenAPI spec version: 0.1.0
 */

export interface PairingStatus {
  connected: boolean;
  /** @nullable */
  phoneNumber: string | null;
  /** @nullable */
  sessionId: string | null;
  /** @nullable */
  connectedAt: string | null;
  /** @nullable */
  uptimeSeconds: number | null;
  /** @nullable */
  lastSeen: string | null;
}